fix(relay-core): split shadowtls version tests under LoC budget
The version-mismatch end-to-end coverage pushed tests.rs to 1513 counted
lines (cap 1500), failing check_file_loc_limits.py (gradle-static-analysis
and architecture-health). Move shadowtls_inner_fixture + the two version-
mismatch tests into a tests/shadowtls_version.rs submodule (use super::*).

Also bump the async-safety allowlist entry for ripdpi-shadowtls client.rs
to_socket_addrs from line 33 to 34: a new import shifted the documented-safe
call down one line, which check_async_safety.py keys on exactly.
